Fix some more windows compile error

+++ PlayMode.hpp
@@ -23,14 +23,14 @@ struct PlayMode : Mode
 
     //----- game state -----
     Scene scene;
-    enum Fruit : uint8_t
+    enum Fruit : int
     {
         CHERRY = 0,     // normal speed, normal acc, low cooldown
         WATERMELON = 1, // low speed, high acc, high cooldown
         ORANGE = 2,     // high speed, no acc, normal cooldown
         BANANA = 3      // normal speed, low acc, normal cooldown
     };
-    uint8_t current_fruit = CHERRY;
+    int current_fruit = CHERRY;
     uint64_t fruit_count = 0;
     float drop_timer;
     static const int COUNTDOWN = 60;
