+++ Scene.hpp
@@ -17,7 +17,7 @@ struct GameObject
     float pos_x;
     float pos_y;
     glm::vec2 velocity = {0, 0};
-    uint8_t extra = -1;
+    int extra = -1;
 
     struct SpriteInput
     {
